requirements.txt
Django==4.0.6
djangorestframework==3.14.0
djangorestframework_simplejwt==5.2.0
django-cors-headers==3.11.0
django-seed==0.3.1
Faker==15.0.0
psycopg2-binary==2.8.6
